Noted change of handle
I recently changed my GitHub handle from @jdbartlett to @redoPop and created this account and fork to make sure I don't break any prior references to my code. The canonical version of this repository is now:

http://github.com/redoPop/innershiv

Should the need arise to reopen this project (heaven forbid!) it will be that version I reopen, and not this fork.